+++ server/src/main/kotlin/it/polito/wa2/g29/server/model/Expert.kt
@@ -10,4 +10,4 @@ class Expert(
     var surname: String,
     var email: String,
     // var expertises: Set<Expertise> = setOf()
-) : EntityBase<Long>()
+) : EntityBase<Int>()
